Migrate reactor event listener
