+++ gravitee-apim-gateway/gravitee-apim-gateway-reactor/src/main/java/io/gravitee/gateway/reactor/impl/DefaultReactor.java
@@ -15,19 +15,14 @@
  */
 package io.gravitee.gateway.reactor.impl;
 
-import io.gravitee.common.event.Event;
-import io.gravitee.common.event.EventListener;
 import io.gravitee.common.event.EventManager;
-import io.gravitee.common.service.AbstractService;
 import io.gravitee.gateway.api.ExecutionContext;
 import io.gravitee.gateway.api.Request;
 import io.gravitee.gateway.api.Response;
 import io.gravitee.gateway.api.context.SimpleExecutionContext;
 import io.gravitee.gateway.api.handler.Handler;
 import io.gravitee.gateway.env.GatewayConfiguration;
-import io.gravitee.gateway.reactor.Reactable;
 import io.gravitee.gateway.reactor.Reactor;
-import io.gravitee.gateway.reactor.ReactorEvent;
 import io.gravitee.gateway.reactor.handler.EntrypointResolver;
 import io.gravitee.gateway.reactor.handler.HandlerEntrypoint;
 import io.gravitee.gateway.reactor.handler.ReactorHandlerRegistry;
@@ -42,7 +37,7 @@ import org.springframework.beans.factory.annotation.Qualifier;
  * @author David BRASSELY (david.brassely at graviteesource.com)
  * @author GraviteeSource Team
  */
-public class DefaultReactor extends AbstractService<Reactor> implements Reactor, EventListener<ReactorEvent, Reactable> {
+public class DefaultReactor implements Reactor {
 
     private final Logger LOGGER = LoggerFactory.getLogger(DefaultReactor.class);
 
@@ -118,34 +113,5 @@ public class DefaultReactor extends AbstractService<Reactor> implements Reactor,
 
     private void processResponse(ExecutionContext context, Handler<ExecutionContext> handler) {
         responseProcessorChainFactory.create().handler(handler).handle(context);
-    }
-
-    @Override
-    public void onEvent(Event<ReactorEvent, Reactable> event) {
-        switch (event.type()) {
-            case DEPLOY:
-                reactorHandlerRegistry.create(event.content());
-                break;
-            case UPDATE:
-                reactorHandlerRegistry.update(event.content());
-                break;
-            case UNDEPLOY:
-                reactorHandlerRegistry.remove(event.content());
-                break;
-        }
-    }
-
-    @Override
-    protected void doStart() throws Exception {
-        super.doStart();
-
-        eventManager.subscribeForEvents(this, ReactorEvent.class);
-    }
-
-    @Override
-    protected void doStop() throws Exception {
-        super.doStop();
-
-        reactorHandlerRegistry.clear();
     }
 }
