+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/rest/api/service/v4/impl/ApiNotificationServiceImpl.java
@@ -67,7 +67,6 @@ public class ApiNotificationServiceImpl extends AbstractService implements ApiNo
 
     private void triggerNotification(final ExecutionContext executionContext, final ApiHook hook, final IndexableApi indexableApi) {
         String userId = getAuthenticatedUsername();
-        System.out.println(userId);
         if (userId != null && !getAuthenticatedUser().isSystem()) {
             UserEntity userEntity = userService.findById(executionContext, userId);
             notifierService.trigger(
