+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/test/java/io/gravitee/rest/api/service/impl/ApiDuplicatorService_RecalculateApiDefinitionIdsTest.java
@@ -53,6 +53,8 @@ public class ApiDuplicatorService_RecalculateApiDefinitionIdsTest {
 
     private static final ObjectMapper mapper = new ObjectMapper();
 
+    private static final ObjectMapper mapper = new ObjectMapper();
+
     @InjectMocks
     protected ApiDuplicatorServiceImpl apiDuplicatorService;
