+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/test/java/io/gravitee/rest/api/service/v4/impl/ApiServiceImplTest.java
@@ -322,11 +322,13 @@ public class ApiServiceImplTest {
     @Test
     public void shouldCreateWithListener() throws TechnicalException {
         when(apiRepository.create(any()))
-            .thenAnswer(invocation -> {
-                Api api = invocation.getArgument(0);
-                api.setId(API_ID);
-                return api;
-            });
+            .thenAnswer(
+                invocation -> {
+                    Api api = invocation.getArgument(0);
+                    api.setId(API_ID);
+                    return api;
+                }
+            );
         NewApiEntity newApiEntity = new NewApiEntity();
         newApiEntity.setName(API_NAME);
         newApiEntity.setApiVersion("v1");
@@ -369,9 +371,10 @@ public class ApiServiceImplTest {
         verify(apiMetadataService, times(1))
             .create(
                 eq(GraviteeContext.getExecutionContext()),
-                argThat(newApiMetadataEntity ->
-                    newApiMetadataEntity.getFormat().equals(MetadataFormat.MAIL) &&
-                    newApiMetadataEntity.getName().equals(DefaultMetadataUpgrader.METADATA_EMAIL_SUPPORT_KEY)
+                argThat(
+                    newApiMetadataEntity ->
+                        newApiMetadataEntity.getFormat().equals(MetadataFormat.MAIL) &&
+                        newApiMetadataEntity.getName().equals(DefaultMetadataUpgrader.METADATA_EMAIL_SUPPORT_KEY)
                 )
             );
         verify(membershipService, times(1))
@@ -386,11 +389,13 @@ public class ApiServiceImplTest {
     @Test
     public void shouldCreateWithListenerAndFlows() throws TechnicalException {
         when(apiRepository.create(any()))
-            .thenAnswer(invocation -> {
-                Api api = invocation.getArgument(0);
-                api.setId(API_ID);
-                return api;
-            });
+            .thenAnswer(
+                invocation -> {
+                    Api api = invocation.getArgument(0);
+                    api.setId(API_ID);
+                    return api;
+                }
+            );
         NewApiEntity newApiEntity = new NewApiEntity();
         newApiEntity.setName(API_NAME);
         newApiEntity.setApiVersion("v1");
@@ -436,9 +441,10 @@ public class ApiServiceImplTest {
         verify(apiMetadataService, times(1))
             .create(
                 eq(GraviteeContext.getExecutionContext()),
-                argThat(newApiMetadataEntity ->
-                    newApiMetadataEntity.getFormat().equals(MetadataFormat.MAIL) &&
-                    newApiMetadataEntity.getName().equals(DefaultMetadataUpgrader.METADATA_EMAIL_SUPPORT_KEY)
+                argThat(
+                    newApiMetadataEntity ->
+                        newApiMetadataEntity.getFormat().equals(MetadataFormat.MAIL) &&
+                        newApiMetadataEntity.getName().equals(DefaultMetadataUpgrader.METADATA_EMAIL_SUPPORT_KEY)
                 )
             );
         verify(membershipService, times(1))
