+++ pom.xml
@@ -60,7 +60,7 @@
         <gravitee-connector-api.version>1.1.2</gravitee-connector-api.version>
         <gravitee-expression-language.version>1.10.1</gravitee-expression-language.version>
         <gravitee-fetcher-api.version>1.4.0</gravitee-fetcher-api.version>
-        <gravitee-gateway-api.version>1.37.3</gravitee-gateway-api.version>
+        <gravitee-gateway-api.version>1.38.0-SNAPSHOT</gravitee-gateway-api.version>
         <gravitee-license-node.version>1.3.1</gravitee-license-node.version>
         <gravitee-node.version>1.25.0</gravitee-node.version>
         <gravitee-notifier-api.version>1.4.1</gravitee-notifier-api.version>
